+++ Sources/PermissionsSwiftUI/Modifiers/Public/MainTextModifiers.swift
