+++ Sources/PermissionsSwiftUI/SwiftUI/Modal-style/ModalView.swift
