+++ Sources/PermissionsSwiftUI/SwiftUI/Shared/AllowButtonSection.swift
