+++ Sources/PermissionsSwiftUI/SwiftUI/Shared/ExitButtonSection.swift
