+++ Sources/PermissionsSwiftUI/SwiftUI/Shared/HeaderView.swift
